define basic new theme
