+++ website/src/app/[lang]/[region]/v2/(home)/(sections)/explainer-video.tsx
