+++ website/src/app/[lang]/[region]/v2/(home)/(sections)/mobile-illustration.tsx
