+++ website/src/app/[lang]/[region]/v2/(home)/(sections)/monthly-income.tsx
